# docs/openapi3_0_prod.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Produto API doc",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/v1/{id}": {
      "get": {
        "tags": [
          "Produto"
        ],
        "summary": "Realiza a busca de um produto por id",
        "operationId": "findById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Produto"
        ],
        "summary": "Modifica o cadastro de um produto existente",
        "operationId": "atualizaproduto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProdutoRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Cadastro do produto atualizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Parâmetros inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "422": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          },
          "500": {
            "description": "Erro ao atualizar o cadastro do produto",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProdutoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Produto"
        ],
        "summary": "Remove um produto existente",
        "operationId": "deletaproduto",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Produto removido com sucesso"
          },
          "400": {
            "description": "Parâmetros inválidos"
          },
          "422": {
            "description": "Dados inválidos"
          },
          "500": {
            "description": "Erro ao remover o produto"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ProdutoRequest": {
        "required": [
          "descricao",
          "marca",
          "nome"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "marca": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "CABELOS",
              "MAQUIAGEM",
              "CORPO_E_BANHO",
              "UNHAS",
              "ELETRICOS",
              "ESTETICA"
            ]
          }
        }
      },
      "ProdutoResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "nome": {
            "type": "string"
          },
          "descricao": {
            "type": "string"
          },
          "marca": {
            "type": "string"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "CABELOS",
              "MAQUIAGEM",
              "CORPO_E_BANHO",
              "UNHAS",
              "ELETRICOS",
              "ESTETICA"
            ]
          }
        }
      },
...
    }
  }
}